requirements.txt
# Основные зависимости
python-chess==1.999